Support for Lua 5.4+ & C++17 (#111)
* Update checkout action on GHA CI

* Show output on test failure

* Fix test compat with C++17

* Add missing header

* Disable warning in GCC 13

* Fix compatibility with Lua 5.4

The couroutine result index is always 1, not dependent on the number of
results which is not available in < 5.4.
Errors during GC are shown as warnings in 5.4+

* Add Lua 5.4.7 to the test matrix

* Fix compatibility with Lua 5.4.4+

A change in Lua 5.4.4 resets the Lua status to OK before calling the
panic handler.
This affects a test checking for an out-of-memory exception.
Check the message string instead for something memory related which is
"good enough".